xfce4-wmdock-plugin changes
A while ago, Gentoo removed xfce4-wmdock-plugin.
The plugin still worked in Xfce 4.14, but Xfce 4.16 removed support for GTK+2 libraries and plugins, and this plugin no longer works.
Drop it, as well as the old xfconf eclass which it needed (and why it was removed in the first place).

Additionally, xfce4-wmdock-plugin-gtk3 has been moved to xfce4-wmdock-plugin:3 for simplicity and recognition. To migrate:

emerge --deselect xfce4-wmdock-plugin-gtk3
emerge xfce4-wmdock-plugin